Move parsing code to dedicated file
