+++ lib/filter-comments.js
@@ -1,0 +1,3 @@
+export function filterComments(config) {
+  return config.replace(/#[^\n]*/g, '').trim()
+}
